Fix unnecessary enum classes

See merge request tkochi/m88clone2!218

+++ src/pc88/pc88.cpp
@@ -172,14 +172,13 @@ void PC88::TimeSync() {
 //
 void PC88::UpdateScreen(bool refresh) {
   uint32_t dstat = draw_->GetStatus();
-  if (dstat & static_cast<uint32_t>(Draw::Status::kShouldRefresh))
+  if (dstat & Draw::Status::kShouldRefresh)
     refresh = true;
 
   LOADBEGIN("Screen");
 
   if (!screen_updated_ || refresh) {
-    if ((dstat & (static_cast<uint32_t>(Draw::Status::kReadyToDraw) |
-                  static_cast<uint32_t>(Draw::Status::kShouldRefresh))))
+    if (dstat & (Draw::Status::kReadyToDraw | Draw::Status::kShouldRefresh))
     //      if (dstat & (Draw::readytodraw | Draw::shouldrefresh))
     {
       int bpl;
@@ -200,7 +199,7 @@ void PC88::UpdateScreen(bool refresh) {
     }
   }
   LOADEND("Screen");
-  if (draw_->GetStatus() & static_cast<uint32_t>(Draw::Status::kReadyToDraw)) {
+  if (draw_->GetStatus() & Draw::Status::kReadyToDraw) {
     if (screen_updated_) {
       screen_updated_ = false;
       draw_->DrawScreen(region_);
@@ -226,7 +225,7 @@ uint64_t PC88::GetFramePeriodNS() {
 void PC88::Reset() {
   bool cd = false;
   if (IsCDSupported())
-    cd = (static_cast<uint32_t>(base_->GetBasicMode()) & 0x40) != 0;
+    cd = (base_->GetBasicMode() & 0x40) != 0;
 
   base_->SetFDBoot(cd || disk_manager_->GetCurrentDisk(0) >= 0);
   base_->Reset();  // Switch 関係の更新
@@ -250,7 +249,7 @@ void PC88::Reset() {
   else
     opn1_->SetIMask(0x33, 0x02);
 
-  main_iobus_.Out(kPReset, static_cast<uint32_t>(base_->GetBasicMode()));
+  main_iobus_.Out(kPReset, base_->GetBasicMode());
   main_iobus_.Out(0x30, 1);
   main_iobus_.Out(0x30, 0);
   main_iobus_.Out(0x31, 0);
